# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Нууу очень крутая апишка",
        "description": "Документация для REST API Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/v1/auth/password/forgot": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Создать токен для сброса пароля",
                "description": "Генерирует токен для сброса пароля и возвращает его (в тестовых целях).",
                "operationId": "69e3fe54a0bb102642b8c2d3245b47a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ссылка для сброса пароля создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ссылка для сброса пароля успешно создана."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неверные данные"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/v1/auth/password/token-check": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Проверить токен сброса пароля",
                "description": "Проверяет, действителен ли токен, созданный для сброса пароля.",
                "operationId": "d25d91b68d5be0883030f876429be6aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Токен действителен"
                    },
                    "400": {
                        "description": "Неверный или истёкший токен"
                    },
                    "404": {
                        "description": "Запрос не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/auth/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Сбросить пароль пользователя",
                "description": "Позволяет задать новый пароль при наличии действительного токена.",
                "operationId": "a34b01fa6648eadb2aad0e38c6bb36bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль успешно обновлён"
                    },
                    "400": {
                        "description": "Неверный токен или email"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Получить список всех постов блога",
                "operationId": "5c0f0d2a855e6441442286ea4d916405",
                "responses": {
                    "200": {
                        "description": "Список постов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BlogPost"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Создать новый пост",
                "operationId": "45b2d8aa26debfeae380bcae0183d103",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Новый пост"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Контент поста"
                                    },
                                    "footer": {
                                        "type": "string",
                                        "example": "Автор: Алексей"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogPost"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/blog/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Получить пост по ID и увеличить просмотры",
                "operationId": "36603b0908f4a9e09203e78c28ac1957",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о посте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogPost"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "Удалить пост",
                "operationId": "909cb65dda25800948784a86c201aa3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пост удалён"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BlogPost": {
                "title": "BlogPost",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Пример поста"
                    },
                    "body": {
                        "type": "string",
                        "example": "Содержимое поста"
                    },
                    "footer": {
                        "type": "string",
                        "example": "Автор: Алексей"
                    },
                    "views": {
                        "type": "integer",
                        "example": 15
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}